+++ assets/translations/ja.json
@@ -149,8 +149,8 @@
   "maintenanceTitle": "現在メンテナンス中です",
   "maintenanceText": "ご不便をお掛けして申し訳ございませんが、現在TrainLCDサービスはメンテナンス中です。\nメンテナンスの詳細はTrainLCD障害情報ページまたはX(旧Twitter)アカウント@TrainLCDをご確認ください。",
   "openStatusText": "障害情報",
-  "savedRoutes": "保存されたルート",
-  "savedRoutesEmpty": "保存されたルートはありません",
+  "savedRoutes": "保存済み経路",
+  "savedRoutesEmpty": "保存済みの経路はありません",
   "selectFirstStationTitle": "始発駅を選択してください",
   "longPressNotice": "画面を長押しするとメニューが開けます",
   "autoAnnounceBackgroundTitle": "バックグラウンド音声を有効にする",
