+++ src/hooks/index.ts
@@ -11,6 +11,7 @@ export { useCachedInitAnonymousUser } from './useCachedAnonymousUser';
 export { useCanGoForward } from './useCanGoForward';
 export { useCheckStoreVersion } from './useCheckStoreVersion';
 export { useClock } from './useClock';
+export { useCommunityRoutes } from './useCommunityRoutes';
 export { useConnectedLines } from './useConnectedLines';
 export { useConnectivity } from './useConnectivity';
 export { useCurrentLine } from './useCurrentLine';
@@ -47,7 +48,6 @@ export { usePreviousStation } from './usePreviousStation';
 export { useRefreshLeftStations } from './useRefreshLeftStations';
 export { useRefreshStation } from './useRefreshStation';
 export { useResetMainState } from './useResetMainState';
-export { useSavedRoutes } from './useSavedRoutes';
 export { useShouldHideTypeChange } from './useShouldHideTypeChange';
 export { useSimulationMode } from './useSimulationMode';
 export { useSlicedStations } from './useSlicedStations';
