+++ src/index.tsx
@@ -25,6 +25,7 @@ import CustomErrorBoundary from './components/CustomErrorBoundary';
 import TuningSettings from './components/TuningSettings';
 import { queryClient, transport } from './lib/grpc';
 import DeepLinkProvider from './providers/DeepLinkProvider';
+import CommunityRoutesScreen from './screens/CommunityRoutesScreen';
 import FakeStationSettingsScreen from './screens/FakeStationSettingsScreen';
 import PrivacyScreen from './screens/Privacy';
 import RouteSearchScreen from './screens/RouteSearchScreen';
@@ -138,6 +139,12 @@ const App: React.FC = () => {
                           options={options}
                           name="SavedRoutes"
                           component={SavedRoutesScreen}
+                        />
+
+                        <Stack.Screen
+                          options={options}
+                          name="CommunityRoutes"
+                          component={CommunityRoutesScreen}
                         />
 
                         <Stack.Screen
